refactor: 테스트 분리
